bind WinLoss service
